Fixed app and added a button that dosent do anything
