--- AndroidStudio/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Tue Feb 04 16:50:24 MST 2020
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-5.4.1-all.zip
