+++ CS386Project/app/proguard-rules.pro
