+++ CS386Project/app/src/main/res/mipmap-hdpi/ic_launcher.png
