+++ CS386Project/app/src/main/res/mipmap-xxhdpi/ic_launcher.png
