+++ CS386Project/app/src/main/res/values/colors.xml
