+++ CS386Project/app/src/main/res/values/strings.xml
@@ -1,0 +1,6 @@
+<resources>
+    <string name="app_name">CS386Project</string>
+    <string name="action_settings">Settings</string>
+    <string name="the_project_the_game">The Project: the game</string>
+    <string name="startgame">Start Game</string>
+</resources>
