+++ CS386Project/gradle.properties
