+++ Deliverables/D1_Inception.md
@@ -14,6 +14,7 @@ Interviews summary:
 
 Interview 1:
 I interviewed Brian Bingham at 9:30 pm on February 14, 2020. I talked with Brain about our text adventure game. He hasn’t played too many text adventure games but has played a lot of games. He noted the reason he likes playing video games is for the freedom to be creative and explore a new world. He also likes a good story to stay interested in the game. He said, “ Make red dead into a text adventure game ” and he would be interested. The other key point to note is that he likes games to be replayable and branching. He likes to be able to play a game a few times and be able to fully immersive and explore the world. He said that a game in the fantasy realm would be pretty generic but something like sci-fi has a lot of potential. The immersion and exploration we seem to get nailed down but need to work on branching narrative and replayability.
+
 Assessment: Nice to have
